Update patch set 5
Patch Set 5:

The `beta` class disappears from the body tag after this change.

The `beta` class is used by mobile diff and search overlay (not sure if it's supposed to be but it is)
We should review and remove those usecases before doing this.

resources/mobile.special.mobilediff.styles/mobilediff.less

Patch-set: 5