API: Make empty pageprops array associative
Same as T69054, but for an empty pageprops result.

Explicitly set the result array type to associative, so that an empty
array is sent in JSON as a JSON object and not an array.  This removes
the need for complex and fragile client-side handling of multiple data
types for the same key.

Adds a test for both the current change and T69054, which is currently
untested, as well as updating the existing test case for client-side
code that handles the response.

Change-Id: I5dacdaf356bfb4d85a62a7e8972b5e6e14f22458