Update patch set 4
Patch Set 4: Code-Review+1

This works but Im not sure how to handle caching in this case when changing the value. Switching to false isnt instant if something like memcached is running. Its unclear to me how we do this for changes in variables and I would like us to think a little before merging.

Patch-set: 4