Update patch set 4
Patch Set 4:

So after some thinking I figure this is no different from @deviceWidthTablet
Tampering with this config variable is likely to be rare (especially when we disable it it by default) so the non-instant roll out if acceptable.

I'll merge tomorrow but will give some time in case anyone wants to disagree with my assessment.

Patch-set: 4