Fix lint error ModuleLoader defined but not used
Since ModuleLoader is linted it complains that the defined ModuleLoader
is not used anywhere because it is only exposed for testing.

Error: Unused defines from file: resources/mobile.modules/modules.js:
  ModuleLoader

Expose it as a global variable on the mobileFrontend instead

Given that mobile.startup no longer exports a ResourceLoader definition
the test has been updated to remove the test for a non-existence class
in a now non-existent module.

Change-Id: I861f277ade8fe83083c9d1b4dc115b872d190f87