Hygiene: Merge mobile.modules and mobile.oo
MobileFrontend has more modules than it needs to have. If you look at
the modules that are `ready` upon page load in MobileFrontend
there are many. Fragmenting in this way doesn't make sense.

Instead of having multiple modules let's collapse all the common modules
into one single library. This will help with T144402

On the long term this will result in a much smaller and easier to maintain
extension.json

Change-Id: Ief411812a3d2ddc480c8d6429101533482f47e16