Don't constrain SVG thumb requests based on the original file size
SVGs are exempt from the general rule that thumbnails may not be scaled
to dimensions larger than the original image.

This changes the dimension normalizing logic in the case where the
requested thumb height or width (but not both) is undefined, so that
for SVGs, the missing dimension is scaled proportionally to the original
file size.

In practice, for example, https://en.wikipedia.org/w/api.php?action=
mobileview&page=wikipedia&thumbwidth=640&prop=thumb will now return a
640x584px thumbnail, rather than a 103x94px thumbnail.

Bug: T98528
Change-Id: I8fc2a5ae5e47dd29bd1000d02c8522864bd1b04e