Update patch set 1
Patch Set 1:

After switching to pixels, I can see where the issue is. 56+202+375+56+56 are the widths of components in the header in tablet mode. This exceeds 720px so it forces the display to adjust as the container says it can only be 100%. These numbers will need to add up to 720px to be shown at the tablet threshold of 720. The switch to pixels is optional, merging the fix up should help fix the patch youve been working on Baha - what do you think?

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0