Make noresize non-specific to thumbnails and cleanup
It seems that some styles were mixed up a bit. magnify is only used by
thumbnails, yet was in images.less and noresize needs to balance out
responsive max-width on images, which not specific to thumbnails.

Also adapt noresize to not reset the width, but to set max-width
instead, which will be more reliable and match original intent of
T22030.

Updated the comments to refer the phab instead of bugzilla numbers.

Bug: T154432
Change-Id: Id1e46435c02c2fd10f2163717feeb681569b74f9