Allow feature flagging of new header
We will partially roll out the new HTML and styles to
allow testing the impact of converting the search input
to a search icon on small screens (<wgDeviceTabletWidth)
which is now controlled by a new config variable
wgMinervaUseHeaderV2 which is enabled in beta and disabled
in stable by default.

This change makes several minor changes to the interface that
Nirzar has approved and do not need to be feature flagged:
* The new notification icon is deployed
* The search overlay close icon will move from the left to the right
* If logged in with JS disabled you will not be able to search, but given
we hide the login button for JS disabled users and this is a temporary
state this is deemed acceptable
* For cached HTML while header v1 is rolled out, the search button
will have no icon and will spill out the containing container but
it is still usable.

Note: this has been tested on cached HTML and where
wgMinervaUseHeaderV2 is true and where it is false.

Additional changes:
* Browser tests will now run with the new header UI.
* A swat change to the beta cluster should be made to ensure that
build also passes. I'll do this upon merge to master.

Bug: T157987
Change-Id: If1b014feac263ac5632cbbf797f597e43cf74868