Update patch set 7
Patch Set 7:

>I'm confused. Maybe because `info` is a broad term, or maybe because footerlinks doesn't have a well defined shape. I'd appreciate another reviewer's views on this patch.

It is confusing but this is the structure that is used to power the desktop footer. :/

If we are not prepared to take the approach in PS1 and have hacks in our code, we will need to behave more like the desktop code.

Alternatively we could of course refactor the footer code in core but I think that would be a great increase in scope for a bug fix!

Patch-set: 7