Fix vertical alignment of notifications label
Make the circle a fixed size like the notifications font size, based on
the zeplin mockups (24px).

Center the text using flex box instead of the mix of line-height and
vertical align that behave weirdly in chrome mobile.

Safari <= 8 will show the number 1px down misaligned, this patch trades
the fix for all for a slight misalignment for browsers without flexbox
support.

Tested on:
* Chrome Mobile 56: https://phabricator.wikimedia.org/F5917091
* Chrome 56: https://phabricator.wikimedia.org/F5917045
* Firefox 51: https://phabricator.wikimedia.org/F5917044
* Safari 9: https://phabricator.wikimedia.org/F5917046
* iOS 8, Safari 8: https://phabricator.wikimedia.org/F5917043

Bug: T158957
Change-Id: I3aa1beb9b12d99f4b08dd7f98c4455156a0dc458