diff: $linkInfo is not an array of query parameters
Follows-up 7953b583b. $linkInfo only contains one item now and it happens to
match the query parameter for the use case needed in InlineDifferenceEngine
for action=markaspatrolled. However, getMarkPatrolledLinkInfo() has no
association with that Action subclass. It's return value is not
"query parameters to make a request that will patrol the change" (if it was,
it would set 'action' as well).

It's a generic data array that used to contain other keys, too - and may
again in the future. Avoid breakage when that happens by not relying on this
undocumented coincidence. This makes the code similar to how it was before.

Bug: T151837
Change-Id: Ie537e850a2333517cbb81231a1df996b07888cb6