Update patch set 8
Patch Set 8: Verified-1 Code-Review-1

Need some help here, system message somehow isn't shown correctly. Just the <this-message-does-not-exist> form is shown and I don't have a clue why. In addition this reuses a lot of code of the from the functionality to show the latest edit in a bar at the bottom of mobile view. Duplicated code maybe should be avoided somehow.

Patch-set: 8
Label: Verified=-1
Label: Code-Review=-1