Update patch set 1
Patch Set 1:

The mobile URL of what? The point is that the URL doesn't serve anything to certain clients.

The special page is not real content, so the canonical URL should be something with equivalent functionality which can be considered the master URL. The actual wiki page is the location most people can make sense of which includes the same information, hence satisfying the definition: "content that is either duplicative or a superset of the content" https://tools.ietf.org/html/rfc6596 .

Serving an HTTP code > 400 is better than nothing though.

Patch-set: 1