Update patch set 1
Patch Set 1:

The mobile site does have content however and you are changing the canonical url there too. The canonical url however points to a page that 404s. I'd rather we didn't hack around yet another thing in core especially when the pages you are linking to are not the same.

An article already points to the desktop URL which contains those links.

If you are so keen to avoid indexing this url why not just add a norobots policy...?

Patch-set: 1