Update patch set 11
Patch Set 11:

As far as I can see this works, at least (I hate this phrase, I should not use this phrase, I hate this phrase) "on my machine" ;) I think I'm going to do a follow up not only for the generalization but also for the style issues mentioned in the task. Those aren't really related to this very bug and I guess splitting those things might keep the history clear.

Patch-set: 11