Update patch set 4
Patch Set 4: Code-Review-1

A removal of 64 lines can not be right when the task lists 63 messages to be removed.

I've gone through the whole en.json. I found two messages which aren't within the result of silver searcher but were removed: mobile-frontend-history-no-results, mobile-frontend-editor-cancel-confirm

I also found one message, that was within the result of silver searcher but was not removed with this patch. That is mobile-frontend-placeholder-app. I couldn't find this one within the code though, so I guess it's already removed. That would mean the silver searcher result has 63 match, one is already removed, means 62 should be removed, and we have 64 removed, and I identified 2 which are too much. The count suggests just keeping those two messages would do it. However, just to be sure, I'm going to test against the qqq.json too.

Patch-set: 4
Label: Code-Review=-1