Update patch set 1
Patch Set 1:

(3 comments)

> I don't get your first line. The canonical URL should be one that returns HTTP 200. Why would the full url of the article return 404?

That's not what I'm saying.
I'm saying the current canonical url to the desktop equivalent 404s thanks to your earlier patch.
A browser trying to index this page will follow the canonical link to desktop and if it behaves should assume that it's not indexable.

As far as I can see the matter is resolved and this adds technical debt for no clear purpose.
You are wrestling with the canonical URL as defined by core and replacing it with something that is not useful and if anything could negatively impact SEO. You are pointing to the canonical desktop URL of the article which is not the same thing and far worse might lead to a robot thinking the mobile version of the languages page is the mobile friendly version of the article...

I'm pointing out that a mobile article already has a canonical URL to the desktop article which has the languages in the HTML already so there is no need to index this page in the first place.

Patch-set: 1