Login and logout url should always be relative
Trust core to enforce https.
Login is always on same domain so URL does not need to be an
absolute url.

This dates back to the days where mobile enforced https
and desktop did not. Now that desktop and mobile are consistent
there is no need to need special case this.

Change-Id: I630176f5bb0fdb65b6279253c42f3ebee7bac709