Update patch set 1
Patch Set 1:

It's on my list of things to document in some kind of "best practices" guide. But for now, the main purpose is to make MediaWiki initialization faster/simpler. For example, let's say your hook modifies the toolbox or something, and you run it if some conditional is true. If the hook is registered conditionally, that condition is run for every MW request, including those like api.php and load.php which don't even have a toolbox, and running the condition was needless. If the condition is simple, it might not make a real difference, but that's not always the case in extension's I've surveyed. And keeping everything in extension.json makes auditing a lot easier, etc.

Patch-set: 1