Update patch set 4
Patch Set 4:

Just advice: It may be useful to instead move away from it being global in the first place. Aside from DOM events (which should be bound on document or window directly), anything else surely can belong to an actual object instead of an anonymous global one?

For example, the events emitted by Skin, could be emitted _on_ the Skin itself (by mixing in EventEmitter). The only inconvenience that this introduces is that if there is an on() listener outside the skin module, that module must have a reference to the Skin object. if that is the case, that may in itself uncover debt that may be worth exploring first. If not, or if a reference to Skin is no problem, then making an emitter might be a better solution for you.

Patch-set: 4