Update patch set 1
Patch Set 1:

Is the goal of the function to prevent lazy loading of images while in the print mode? If so, images don't load without the patch on Chromium Version 57.0.2987.133, for example. Can you share instructions on how to test this patch? I wonder if an alternative approach of hiding image placeholders in print mode does the job better? This way you can easily match un-hidden placeholders and load images for them.

Also, the linked task doesn't seem to be related to this patch.

Patch-set: 1
Label: Verified=0