Update patch set 1
Patch Set 1:

The linked task is related. Olga asked to stop lazy loading images for symbols and Math because she didn't think it was possible to stop lazy loading everything (a good example of why tasks should be generic and explain problems not solutions)

The goal of the function is to prevent lazy loading of images while in print mode for browsers that support it.

> If so, images don't load without the patch on Chromium Version 57.0.2987.133, for example
What does this sentence mean?

"Without the patch" of course. This is the problem that it tries to fix.

To verify the fix,
1) before applying the patch open a page with images in Chrome and try to print a page. Notice how in the print version the images are not present
2) Apply the patch and open a page with images in Chrome  and notice how the images are lazily loaded, but when print is clicked all the images are present in the printed document.

I'm not sure I understand your alternative solution. There are no images in the HTML to show or hide. The only images in the HTML are wrapped in noscript tags which cannot be revealed by CSS alone.

Patch-set: 1