Update patch set 2
Patch Set 2:

> 2) Apply the patch and open a page with images in Chrome  and notice how the images are lazily loaded, but when print is clicked all the images are present in the printed document.

So the goal of the patch is to load all images before printing? The commit message is confusing then.

I've followed the above instructions and couldn't get the printed PDF to include images that haven't been lazy-loaded. To clarify, I opened a page where there were two images. The first image was at the top of the page, so it was loaded immediately. The second image was not loaded as it was towards the bottom of the page. Then I printed the page to PDF and didn't see the second image in the printed PDF.

Patch-set: 2