Update patch set 3
Patch Set 3:

> Unlike the mock-ups the table rows have double horizontal borders - border bottom from the previous row and border top of the current row.

I'm not sure I understand the difference between having a 2px border top and 1px of each. I did so to avoid problems with the borders at the bottom of the table (given the table already has a border). Can you elaborate?

> Vertical padding also seems lower than what the mock-ups have.
The article title at the top of the infobox should be italic.
'm testing the article "Bukhara" and the item names (in left column) are aligned in the center. They should be left-aligned according to the mocks.
There should not be a border between the title and the image inside the infobox.
Links should not be underlined.
Some bullet points have bullets in front of text. Those bullets should be removed.

So, I wanted to talk to Nirzar about styling the infobox. No infobox follows the exact same pattern and I'm not convinced we can apply styles to all infobox and consistently make them look good. Infoboxes tend to include inline styles so we'd end up wrestling and trying to override quite a few of them (for instance this one has a background - do we override that too? https://de.m.wikipedia.org/wiki/Weenermoor ).

I think a good starting point would be to at least have a shared sample of infoboxes in test articles using different templates (note there are hundreds of infobox templates :)).

I get uneasy about attempting to style HTML we do not control so I think this part of the print design is going to require a lot more thought...

I'll setup a meeting.

Patch-set: 3
Label: Verified=0