Improve layout of fallback editor
This attempts to make the save changes button visible without
the user having to scroll down the page and clean up the UI.
Note for anonymous users, the user will have to scroll as the warning
text appears before the form. This is unavoidable and an encouragement
to login!

Styles apply when show changes or show preview have been clicked

When the non-JavaScript editor is open
* Reduce heading font-size
* Reduce size of checkbox labels, warnings and licenses to be consistent
with license box in overlays. Pull out a generic mixin until this
component can be generalised. Note given
skins.minerva.fallbackeditor is only applied when action=edit
there is no additional bloat to the resulting CSS payload on standard
mobile views
* Reducing padding of elements to make room for the important textarea
and buttons.
* In preview mode, style "Remember this is only a preview" as a warningbox
to improve visual separation from rest of content
* Remove use of large headings in editor which make UI more cluttered

Bug: T125174
Change-Id: Ibb7355b828c3bedbe8ba9aeaf95a0391c42a18d6