Update patch set 4
Patch Set 4:

I'm not convinced, looks bit hack-ish. First we added new hook and now we try to make it work with new hook. I would prefer to keep the SkinMinerva::getContextSpecificModules() as it was before and move the setting Skin::setSkinOptions() somewhere earlier than onBeforePageDisplayMobile(). Probably the only possible option is "ExtensionFunctions" but it is to early as we do not have Skin nor Config objects initialized yet

Patch-set: 4