QA: Make sure to use correct selector
In code we treat #searchIcon as the search icon, and not
the input element inside it. Make sure tests use the same
element for the search icon.

This change should prevent browsers from submitting the header
form. The browser should instead trigger opening of the search
overlay. As a result we should be able to get rid of intermittent
browser test failures of the scenario "Search with search in
pages button".

Bug: T155855
Change-Id: I27f70bc9371219083e8a3c88f80e556dc3f0feba