Please don't abuse global events on M object when you dont need to
The SearchOverlay makes use of the global event emitter, apparently
just so it can log events. This is completely unnecessary as the
search logger is created at the same time as the search overlay.
Instead of doing this use the fact that a SearchOverlay has an
EventEmitter mixin and use those events instead.

If we want to use global events we should probably be using
mw.track instead as it is supported by core.

Change-Id: I862114e59e1b7816bfc5b1bfd31e703bda463d91