Search button should not be wrapped in a link
This looks like an artifact of the old search form. The problem with
using both a link and a submit icon, is that in certain browsers
the link may be followed
rather than the submit.

Clicking the search icon submit button will have the same effect as
clicking the link - it will submit with an empty search query.

There is however no need to include the submit button.
Without JS, the link will take the user to the search form.

With JS, hitting enter is what submits the form which happens
even if it is omitted.

Also add method to form to avoid ambiguity.

Bug: T140490
Bug: T163049
Change-Id: Ide7725dff259596c963542a0171a78f38ed7e038