Update patch set 3
Patch Set 3:

(1 comment)

In that comment I was using the term "animate" to refer to a CSS animation - in contrast to a transition. Where a transition is used to "transition" between two values, an animation is used to unconditionally perform an animation directly when an element first matches a particular rule.

In current master, we have a placeholder that reserves a white block. When it nears the viewport, we create a detached image, wait for it to load, and when it finishes we append it to DOM, and add a class to the placeholder to remove the white reservation - which changes through a transition instead of applying immediately.

What I was suggesting is to, when the placeholder nears the viewport, immediately replaces it with the image in question and apply the class. This way it loads natively into the document instead of detached. The class can then trigger an animation instead of a transition. From grey to transparent, like before, and for the same duration. The benefit is that if the image loading takes longer than the transition, the image will load directly (like it would normally for non-lazy loaded images) into the viewport progressively or interfaced - which is arguably a better user experience than holding up a box with a solid color until the image is done (which may take a long time, or fail with a partial render that we then never show the user..).

The animation also supports a delay to at least allow the fade to happen completely for the common case (e.g. 200ms or so and then unconditionally start the fade).

I recommend an animation instead of a transition because it is capable of tracking the delay natively instead of in JavaScript, otherwise you're dependant on load events or timers to trigger a class chance, which is a problem because they can only happen if the browser is giving you a slot to execute code. When the user is scrolling, or the browser is otherwise busy, these slots are throttled and the image may be fully dowloaded for a long time but you'll never see it because Skin.js is waiting for its turn to execute code and swap the class.

This is currently a problem already and I've pointed this out early on various tasks as well when we discussed the use of the fade before, at which point I said we can keep the fade if desirable, but per the above, it does not need to come with the compromise of keeping the image hidden longer than needed.

Patch-set: 3
Label: Verified=0