Update patch set 3
Patch Set 3:

(1 comment)

> In that comment I was using the term "animate" to refer to a CSS animation - in contrast to a transition. Where a transition is used to "transition" between two values, an animation is used to unconditionally perform an animation directly when an element first matches a particular rule.

This is what confused me.. we are currently using animations via @keyframe. We are not using a transition. Did you mean we drop the animation of the container and just animate the image?

> In current master, we have a placeholder that reserves a white block. 

I think you mean gray?

> The benefit is that if the image loading takes longer than the transition, the image will load directly (like it would normally for non-lazy loaded images) into the viewport progressively or interfaced - which is arguably a better user experience than holding up a box with a solid color until the image is done (which may take a long time, or fail with a partial render that we then never show the user..).
The animation also supports a delay to at least allow the fade to happen completely for the common case (e.g. 200ms or so and then unconditionally start the fade).

I believe what you mean is this https://gerrit.wikimedia.org/r/349717 ? "Arguably" is the keyword here. This is up to Nirzar, but he has expressed a strong preference to only reveal the image when it's fully available.  I should note Medium in their example (see https://jmperezperez.com/medium-image-progressive-loading-placeholder/) add the class to the onload event using css transitions which has a similar effect. Maybe this is a compromise?

> I recommend an animation instead of a transition because it is capable of tracking the delay natively instead of in JavaScript, otherwise you're dependant on load events or timers to trigger a class chance, which is a problem because they can only happen if the browser is giving you a slot to execute code. 

Sure no misunderstanding here, which is why I threw up this patch as I thought you were suggesting using $.animate which seemed strange to me.

Patch-set: 3
Label: Verified=0