Update patch set 1
Patch Set 1:

With this approach, you'll want to set a reasonable delay on the animation so that it doesn't start fading immediately. Currently this will 100% of the time start fading when the image isn't there yet, which means the first parts of the image will always appear without fade.

A delay between 300ms and (1s-0.3s) 700ms seems reasonable given RAIL expectations. After that delay if onload still hasn't happened, the fade would just start on its own and from that point forward we let the streamed download and render complete natively in whatever way the browser normally would.

From Navigation Timing:
* responseStart (TTFB) p50 is 200ms, p75 is 500ms, p95 is 2s.
* responseEnd is typically responseStart +100ms.

The idea is to have UX and design remain exactly as it is today (as approved by Nirzar previously): a 0.3s fade that starts as soon as the image download completes fully. With the exception that:

1. If the image download completes normally but JS isn't getting execution time (e.g. due to scroll throttling or other expensive tasks the browser is doing, especially common on mobile), the fade will start on its own eventually instead of only after 1-2 seconds when the user stops scrolling or when JS execution is unlocked.

2. If the download takes more than 2 seconds, the fade will happen with whatever has been rendered internally so far (possibly nothing, in which case the fade is invisible to the user) and the image download/render is completed natively further, friction-free.

Here's a quick mock up of the end result (ignore the code): https://codepen.io/Krinkle/full/VbjKEo

Patch-set: 1