Category overlay should show "content based" categories
Only hidden categories should appear in "organizational"
Now we are using formatversion=2 rather than checking for empty string
we'll want to use the boolean

Browser tests have been added... but they will only run on the beta cluster.
Category updates are done via the job queue so it is not possible to create
a page with categories within an @integration test.

Bug: T163699
Change-Id: Ie49bac6b9152cfd9ef0f67399d021334198cef13