Update patch set 1
Patch Set 1:

Indeed, though not because of IE 9, which no longer receives JavaScript. Typical use will be like this:

 var rAF = window.requestAnimationFrame | mw.requestIdleCallback;

You can see this in various places in core, Vector and VisualEditor. Typically for either 1) reading style after computation, or 2) preloading a variable that holds style information before it is needed, but also not doing it synchronously on page load, 3) edge cases where class additions should be processed separately instead of together to achieve the desired transition or animation.

Patch-set: 1