Update patch set 10
Patch Set 10:

Given the issue with reflows that is not solved by https://gerrit.wikimedia.org/r/#/c/349275/5 and the fact we may want to lazy load inline images for printed copies of the mobile site, I think this is a useful config option to have available.

It's disabled by default so will not impact production.

Patch-set: 10