Fix invalid state error
MobileFormatter::makeSections() was modifying the DOM structure
and because of that MobileFormatter::makeHeadingsEditable() was failing
as subheadings were rearranged (wrapped into different containers).

MobileFormatter::makeHeadingsEditable() doesn't change the DOM structure
so flipping those two calls solved the problem. First we add classes to
existing elements, then we change the DOM.

Bug: T151838
Change-Id: I750bcefc63c11164b5eb5f9c6c1a209355d8dbab