Update patch set 3
Patch Set 3: -Code-Review

Looks like the image being rendered has a width of 350px and height of 233px so the placeholder has a width of 350px and height 233px.

When the image is loaded a CSS rule max-width: 100% rule kicks in to downsize it to fit into the constraints of the skin causing the reflow - 288px by 192px

I'm not sure how we can avoid this. We can obviously add overflow:hidden+display:block to all .image's but this only solves the issue of horizontal reflow.. not vertical.

This seems to be the case with display: block and display: inline-block so I think the original bug fix was ill-informed, but it probably explains the remaining reflows we are seeing in T146298

The more general issue here is that we have to apply max-width: 100%; height: auto;
Ideally all thumbnails would be standardised to a mobile friendly width/height. We can do that in MobileFormatter.

What thoughts do you have @krinkle

I think this patch is fine, but there are still clearly some pending reflows we need to work out how to solve.

Patch-set: 3
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0