Update patch set 1
Patch Set 1:

> <img> supports background-color without issue.
Example: https://codepen.io/Krinkle/pen/xdEmrp

Sure.. but as soon as the image starts to load the background disappears.. at least for me in Chrome.

In the codepen you linked to the animation is totally different experience from the current animation.

Try it with a large image for example:
https://upload.wikimedia.org/wikipedia/commons/thumb/f/f6/Panel_fields_questions_on_Journalism_and_Wikipedia_%28cropped%29.jpg/1650px-Panel_fields_questions_on_Journalism_and_Wikipedia_%28cropped%29.jpg

I don't see any animation this way - as soon as the src is added the image is loaded the way the browser decides:

https://phabricator.wikimedia.org/F7783171

Patch-set: 1