Document removeXMLProcInst line in .svgo.yml
If your SVG doesn't start with an XML declaration, then it's MIME type
will be detected as "text/plain" rather than "image/svg+xml". This is
because CSSMin::getMimeType uses the fileinfo extension to detect the
MIME type of a file and that extension to defers to libmagic. libmagic's
default database[1] currently requires that SVGs contain an XML
declaration.

1: https://github.com/threatstack/libmagic/blob/master/magic/Magdir/sgml#L5

Change-Id: Ibacb748fec161f08a2e2d9ca15847da0e3cc8ccd