Update patch set 5
Patch Set 5: Code-Review-1

I've tested locally and it does some weird things to me.

I login with a user that has a notifo, and the red circle with a 1 swaps to a red circle with a 0. Doesn't happen in master.

See screen cap i made https://gfycat.com/AfraidUnlawfulBrownbear

Patch-set: 5
Label: Code-Review=-1