All dd elements should allow scrolling
If a dd element contains an image the image may extend out of
the viewport. To protect against this we make it overflow scroll.
This will not impact any text inside the dd and dd tag is already
display block so will not result in any UI regressions

Using overflow: auto will show a scrollbar if needed but will not
show one if it is not. Using overflow: scroll would make all dd elements
show a scrollbar even when not necessary.

Bug: T160946
Change-Id: I796a023e497a47655482bf35df85d2ff030c3dc6