Hygiene: Fix View.js code examples
'@example' is reserved for a code block that will produce an
interactive example. MobileFrontend doesn't currenlty have
a jsduck eg-iframe configured, so these examples are broken.

In addition, it also fails with `SyntaxError: Unexpected token <`

https://doc.wikimedia.org/MobileFrontend/master/js/#!/api/View

Instead, for example code that is merely meant to be a syntax-highlighted
block, Markdown indentation will do the trick.

Change-Id: Iba52966c1c4ed24d1a13637f3740fa80801f6754