Not-existent references should fall back to hash change
Several deprecated templates point to hash fragments that are not
actually references and could refer to headings/sections/notes at the end
of sectons.

Instead of throwing an error here or getting stuck in an infinite loop
navigate to the hash fragment. Do not try and guess the behaviour in this
circumstance and render the HTML that it refers to.

Additional changes:
* getReference now rejects if no reference is found

Bug: T154861
Change-Id: I7434d9b0bcdbdc7a076db5127ed62866019ce99b