Correctly handle the mw-parser-output wrapper
Since If4eb5bf71f94fa36, MW parser output intended for the primary page
view will have a <div class="mw-parser-output"> wrapper around it. So
look for headings as children of that wrapper, instead of at the top
level of the document.

Renamed $sibling to $containerChild, since it didn't make sense to me.
Sibling of what? It could be the only child.

Tested with old and new MW core.

Bug: T164733
Change-Id: I52fe218c6ffe082ec3d5c79167c24105c828eeca