Update patch set 3
Patch Set 3:

There is one minor issue that should be solved. Currently MobileFormatter is always looking for a "mw-parser-output" class, but there is a "ParserOptions::setWrapOutputClass()" that can override "mw-parser-output" to something else. The classname can be pass as an API param (https://gerrit.wikimedia.org/r/#/c/350634/4/RELEASE-NOTES-1.30)


If someone uses different class this solution won't work. For now it's ok but IMHO we should tackle that so the whole codebase is consistent instead of hardcoding values that might change.

Patch-set: 3