Update patch set 1
Patch Set 1: Code-Review-1

Can you raise a bug?
This appears to be using the noresize class on the `img` tag when it should be doing so on the `noresize` element. It can and should be fixed on wiki.

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1