Only wrap top level headings
Previously we naively assumed any headings in the page were
headings that could be used to collapse sections that followed
however our editors use headings in other ways - e.g. on portal
pages the headings correspond to text in boxes.

Revisiting this, it doesn't make sense. Section collapsing only
works if a heading is not wrapped so this change should result
in fewer broken pages.

Changes:
* prepareHeadings is now prepareHeading
* A heading is only transformed when it is encountered during the
DOM traversal
* Correct case of Integer type

Bug: T152055
Change-Id: Id26632fe276c0f8c5720df39137f8a75648a2ba5