Update patch set 15
Patch Set 15: Code-Review-1

Code looks good, but it doesn't work properly.

When you click on the notification count, the overlay opens, then when you close it, the notification count completely disappears, leaving the Bell icon there, instead of the circle with the gray border.

Here's a gif illustrating the issue:

    http://imgur.com/yKqMiZS

Patch-set: 15
Label: Code-Review=-1