rl: Break out mobile.messageBox styles module
The mobile.messageBox styles are loaded using
OutputPage#addModuleStyles, which requires that the specified modules
are styles-only, on certain special pages. They are also loaded by the
RL on other pages as they are depended on by the mobile.nearby and
mobile.editor.common RL modules.

Support both ways of loading the mobile.messageBox styles (and other
content) and the RL by:

* Creating the mobile.messageBox.styles styles-only RL module.
* Making the mobile.messageBox RL module depend on
  mobile.messageBox.styles.
* Only loading the mobile.messageBox.styles RL module on certain pages.

Bug: T164892
Change-Id: I46dd7fbfdfebd1daf732f9c3a7dbbf488ff6c533