rl: Break out mobile.mainMenu style module
The mobile.mainMenu styles are loaded by OutputPage#addModuleStyles,
which requires that the specified modules are styles-only, on the
Special:MobileMenu page. A side effect of this is that the
mobile.mainMenu JS and templates aren't loaded on Special:MobileMenu,
which is fortunate as they aren't required. They are also loaded by the
RL on other pages.

Support both ways of loading the mobile.mainMenu styles (and other
content) and the RL by:

* Creating the mobile.mainMenu.styles styles-only RL module.
* Making the mobile.mainMenu RL module depend on mobile.mainMenu.styles.
* Only the mobile.mainMenu.styles RL module on Special:MobileMenu.

Bug: T164892
Change-Id: Ia108d63753593ca8e6ce266e595d32af148b436a