Display newly added talk topic
Expose PageGateway cache to the module rather than keeping it
private to the instance. The reason is that the various types of
talk overlays, e.g. TalkOverlay, TalkSectionAddOverlay, etc. create
their own instances of PageGateway and a change in one of them didn’t
reflect in the other instances of PageGateway.

Bug: T151041
Change-Id: I026b4a5a7c7fdc13e6115466180cf82d57eed524