Refactor Special:MobileDiff
This makes the code a bit more readable, and reduces the size of
functions by extracting the code to their own functions. Also,
the difference engine is created earlier and is therefore be
available to most of the code.

This change also sets the relevant title of the skin to the title
of the revisions, which are viewed. This way, other extensions now,
that the special page shows content of a (probably) content page,
instead of a real special page.

Bug: T165835
Change-Id: I68cf50f5dd339f34802d70df1f32d2c3390944a3