Introduce skinStyles folder for Minerva's styles
The purpose of the skinStyles folder is to make it more obvious when
we are shipping styles specific to Minerva. This should help with
debugging and it will also help with separating the Minerva skin from the
MobileFrontend component.

Mixing skinStyles with resources can be a little confusing as it is not
obvious that certain files are only loaded by the Minerva skin.

Changes:
* Introduce skinStyles folder for Minerva specific styles that customise
modules that will remain provided by MobileFrontend and other extensions
* Make skins.minerva.notifications.filter.styles a skin style of
 ext.echo.styles.special
* Introduce skinStyles of mobile.startup module and consolidate code previously
shared between the Search, Language overlays. Shift Minerva specific styles
over there.

Change-Id: I1ee00148a248d3725b02eea3c19e6b22486ae5ba