svgo: Update pre-commit hook
Remove the --pretty option as svgo outputs a smaller file without it.
The `dev-scripts/svg_check.sh` script also doesn't have this option.
As a result of this change the remote `npm test` job won't fail if
the job was not run locally before committing.

Thanks to Jon for figuring out the issue.

Bug: T142532
Change-Id: Icf6cb2bf0c97a36bff3888bffc6136c26c871ae1