phpcs: Allow linting tools to discover rules
Explicitly stating the current directory fixes the ambiguity
of finding the ruleset referred to in the file. This allows
syntax checkers like flycheck (for GNU Emacs) to work.

Also ignore node_modules as it contains the `resource-modules` package
which has php source code that doesn't conform to our linting rules.

Change-Id: I4c54e33de16cc3523153d40e8da20f4eee942d2d